Updates to the redirect functionality
Added a configuration option to specify a route to be used to redirect
the user to when they have successfully logged in or out. A default
value is specified in the ModuleOptions.php file. The user will be able
to specify a route in the zfcuser.global.php.dist.

There is another part to this redirect feature which needs some work and
that involves a query string parameter. This way a user could specify a
route for that particular login to over ride the route specified in the
config.