Updates to redirect funcationality
Completed the redirect funcationality. It is now possible to set a
config variable to let a developer specify the route which the user will
be redirected to when they login and logout.

It's also possible to set a ?redirect=<url> query parameter and this
gets passed around which will override the config parameters.