Set identity label in both needed spots
This will set the identity label in both the label attribute, and the label key in the element's options attribute. We are doing that for credential and some packages (zf2-twb-bundle for example) look for the label in the options array instead of the attribute. Best to belt-and-suspenders and have it in both places.